AllLogin_DRF_14_project/backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
cffi==2.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==4.2.24
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
fonttools==4.59.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.10.0
zopfli==0.2.3.post1
